first base-number then kmer counting, working =version, just changed some stream opening that would not compile

+++ src/lib/BarcodeMapping.hpp
@@ -350,7 +350,7 @@ class Mapping : protected MappingPolicy, protected FilePolicy
                               DemultiplexedLine& demultiplexedLine,
                               BarcodePatternPtr pattern,
                               const input& input, 
-                              std::atomic<int>& count, const unsigned long long& totalReadCount,
+                              std::atomic<unsigned long long>& count, const unsigned long long& totalReadCount,
                               int& mmScore,
                               OneLineDemultiplexingStatsPtr stats);
